pyproject.toml
[project]
name = "ecb-2512"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.0",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
]